backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ctomrp <cristian.rico.palma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.8"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
pyjwt = "^2.8.0"
drf-spectacular = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
